!!! BUGFIX: make custom SqlFilters with parameters work; fix Entity Security depending on more than just roles
## Problem Description

It is not allowed to call "setParameter" in the addFilterConstraint method
of an SqlFilter; as this will interfere with Doctrine's query caching:
Doctrine tries to avoid building queries as much as possible. It also
takes the applied SqlFilters *and their parameters* into account when
building cache identifiers, and if it has a matching cache entry, it will
not call "addFilterConstraint". Thus, "setParameter" must be called as early
as possible; way before the "addFilterConstraint" method might be called.

## Visible Error

If you had Entity Constraints or custom SQL Filters which did not only rely
on different user roles, but needed the domain model to build up SQL 
constraints (through entity security or custom SQL Filters), the query
cache in production cached the SQL queries "wrongly", so you were able to
see data from other accounts.

## Bugfix

The bugfix has several parts:

1) Introduce a new `EarlyParameterInitializationSqlFilterInterface`
   to be implemented by SqlFilter which separates parameter
   initialization from query building. Call this interface as early
   as possible, when the security context is ready to be initialized.

2) If you depend on custom globalObjects in Entity Security, ensure
   these can also contribute to the cache identifier by implementing
   CacheAwareInterface.

## Detail Notes

* The error was masked in many cases by making the Doctrine CacheAdapter
  add the security context hash (containing the current roles); so the
  problem only became visible when you used custom globalObjects.

## Breakingness

This change is only breaking if:

- You used Entity Security AND custom context objects registered in
  `TYPO3.Flow.aop.globalObjects`:

  In this case, your global objects should very likely implement
  CacheAwareInterface.

- You used custom SQL Filters:

  In this case, ensure they implement `EarlyParameterInitializationSqlFilterInterface`
  and that there are no `setParameter` calls inside `addFilterConstraint`.